+++ src/Integration/Phpunit/PhpunitMatcherDriver.php
@@ -37,23 +37,33 @@ class PhpunitMatcherDriver extends AbstractMatcherDriver
     }
 
     /**
-     * Get the matcher class name.
+     * Returns true if this matcher driver's classes or interfaces exist.
      *
-     * @return string The matcher class name.
+     * @return boolean True if available.
      */
-    protected function matcherClassName()
+    public function isAvailable()
     {
-        return 'PHPUnit_Framework_Constraint';
+        return class_exists('PHPUnit_Framework_Constraint');
     }
 
     /**
-     * Wrap the supplied matcher in a Phony matcher.
+     * Get the supported matcher class names.
+     *
+     * @return array<string> The matcher class names.
+     */
+    public function matcherClassNames()
+    {
+        return array('PHPUnit_Framework_Constraint');
+    }
+
+    /**
+     * Wrap the supplied third party matcher.
      *
      * @param object $matcher The matcher to wrap.
      *
      * @return MatcherInterface The wrapped matcher.
      */
-    protected function wrapMatcher($matcher)
+    public function wrapMatcher($matcher)
     {
         return new PhpunitMatcher($matcher);
     }