+++ test/suite/Integration/Prophecy/ProphecyMatcherDriverTest.php
@@ -30,28 +30,21 @@ class ProphecyMatcherDriverTest extends PHPUnit_Framework_TestCase
         $this->assertTrue($this->subject->isAvailable());
     }
 
-    public function testIsSupported()
+    public function testMatcherClassNames()
     {
-        $this->assertTrue($this->subject->isSupported($this->matcher));
-        $this->assertFalse($this->subject->isSupported((object) array()));
+        $this->assertSame(array('Prophecy\Argument\Token\TokenInterface'), $this->subject->matcherClassNames());
     }
 
-    public function testAdapt()
+    public function testWrapMatcher()
     {
-        $object = (object) array();
-        $expected = new ProphecyMatcher($this->matcher);
-
-        $this->assertTrue($this->subject->adapt($this->matcher));
-        $this->assertEquals($expected, $this->matcher);
-        $this->assertFalse($this->subject->adapt($object));
+        $this->assertEquals(new ProphecyMatcher($this->matcher), $this->subject->wrapMatcher($this->matcher));
     }
 
-    public function testAdaptWildcard()
+    public function testWrapMatcherWildcard()
     {
         $this->matcher = Argument::cetera();
 
-        $this->assertTrue($this->subject->adapt($this->matcher));
-        $this->assertSame(WildcardMatcher::instance(), $this->matcher);
+        $this->assertSame(WildcardMatcher::instance(), $this->subject->wrapMatcher($this->matcher));
     }
 
     public function testInstance()